move settings button into toolbar
